% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_phyloseq.r
\name{add_tsne_to_sample_data}
\alias{add_tsne_to_sample_data}
\title{Add the results from t-SNE to the phyloseq sample data}
\usage{
add_tsne_to_sample_data(physeq, tsne_obj, prefix = "tsne")
}
\arguments{
\item{physeq}{Phylogenetic sequencing data (\code{\link[phyloseq]{phyloseq-class}}).}

\item{tsne_obj}{An list of class \code{tsne_phyloseq}.}

\item{prefix}{Character string for making new column names.}
}
\value{
A phyloseq object where the t-SNE coordiantes are added to the sample data frame.
}
\description{
Add the results from t-SNE to the phyloseq sample data
}
