% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_distance.r
\name{dist_sne}
\alias{dist_sne}
\alias{dist_to_sne}
\title{Calculate similarity matrices for a t-SNE analysis}
\usage{
dist_sne(xdata, perplexity = 25)

dist_to_sne(dist_, perplexity = 25)
}
\arguments{
\item{xdata}{A matrix or data frame.}

\item{perplexity}{The perplexity. Usually a number between 5 and 50. Default is 25.}

\item{dist_}{A matrix of distances or \code{\link{dist}} object.}
}
\description{
t-SNE uses a similarity measure that can be interpreted as the proability that
two points are nearest neighbors. These functions calculate these probabilities.
}
\details{
\code{dist_sne} computes the SNE similarities using euclidean distances, as in the
originla t-SNE formulation.

\code{dist_to_sne} converts an already computed dissimilarity matrix to a matrix
of SNE similarities.
}
\examples{
data(iris)
iris_distances <- dist(iris[,1:4], method = 'manhattan')
iris_sne_distances <- dist_to_sne(iris_distances)

}
