% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_phyloseq.r
\name{plot_tsne_phyloseq}
\alias{plot_tsne_phyloseq}
\title{Plot microbiota t-SNE results using ggplot2}
\usage{
plot_tsne_phyloseq(physeq, tsne_obj, axes = 1:2, color = NULL,
  shape = NULL, title = NULL, justDF = FALSE)
}
\arguments{
\item{physeq}{Phylogenetic sequencing data (\code{\link[phyloseq]{phyloseq-class}}).}

\item{tsne_obj}{An list of class \code{tsne_phyloseq}.}

\item{axes}{The axes to plot. Must be a length 2 numeric.}

\item{color}{Character string indicating which variable in sample_data should be used as the ggplot2 color aesthetic}

\item{shape}{Character string indicating which variable in sample_data should be used as the ggplot2 shape aesthetic}

\item{title}{Plot title.}

\item{justDF}{If FALSE (default) the plot will be displayed and a ggplot2
object will be returned. If TRUE a data.frame useful will be returned instead.}
}
\description{
This function provides an interface for combining a t-SNE ordination
with sample meta data from a phyloseq object.
}
\details{
This function is fashioned after the \code{\link[phyloseq]{plot_ordination}}
function in the phyloseq package.
}
\examples{
library(phyloseq)
data("GlobalPatterns")
tsne_res <- tsne_phyloseq(GlobalPatterns, distance='wunifrac',
  perplexity = 8, verbose=0, rng_seed = 3901)
plot_tsne_phyloseq(GlobalPatterns, tsne_res,
  color = 'SampleType', title='t-SNE (Weighted UniFrac)')


}
\seealso{
\code{\link{tsne_phyloseq}}. \code{\link{add_tsne_to_sample_data}}.
}
