% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_fit.r
\name{tsne_dist}
\alias{tsne_dist}
\alias{tsne}
\title{Dimension reduction with t-SNE}
\usage{
tsne_dist(dist_, dimensions = 2, verbose = 0, rng_seed = NULL,
  control = list())

tsne(xdata, perplexity = 25, dimensions = 2, verbose = 0,
  rng_seed = NULL, control = list())
}
\arguments{
\item{dist_}{A matrix or \code{dist} object woth pairwise dissimilarities. All elements must be positive and sum to 1.}

\item{dimensions}{The number of dimensions of the t-SNE embedding. Default is 2, which is usually adequate.}

\item{verbose}{The amount of output during the t-SNE estimation. Can be 0 (no output), 1 (some output, including a progress bar) and 2 (detailed output, mostly usefull for debugging purposes)}

\item{rng_seed}{Provide a seed for generating intiial values for the coordiantes. The final t-SNE result is highly sensitive for starting values, so to get reproducible results you should provide a seed.}

\item{control}{A list of control parameters. See 'Details'.}

\item{xdata}{A matrix or data frame.}

\item{perplexity}{The perplexity. Usually a number between 5 and 50. Default is 25.}
}
\value{
A list of class 'tsne' with the following elements:

\tabular{ll}{
\code{par}: \tab A matrix with the t-SNE embedding. \cr
\code{trace}: \tab A vector of the Kullback-Leibler divergences.
If you used early exaggeration (default) you will see a large jump in the
values at iteration 50. \cr
}
}
\description{
t-Distributed Stochastic Neighbor Embedding (t-SNE).
}
\details{
tsne_from_dist assumes that dist_ is a distance matrix interpretable as a
nearest neighbor distribution. If you only have a distance matrix, you can use
the \code{\link{dist_to_sne}} function to create a SNE dissimilarity matrix.

Function \code{tsne} provides a convenient interface to use t-SNE using raw data.
Under the hood it calls the \code{\link{dist_sne}} and \code{\link{tsne_dist}} functions.
}
\section{Control}{

The control argument is a list that can supply any of the following components:

\tabular{ll}{
\code{max_iter}: \tab The number of iterations. Default is 1000. \cr
\code{step_size}: \tab The gradient descent step size. Default is 100, but it can sometimes be usefull to set this to a lower value.\cr
\code{momentum_values}: \tab A vector of values that determines the momentum. The default is c(0.5, 0.8), following the original t-SNE paper.\cr
\code{momentum_iter}: \tab A vector of values that determines the iterations the momentum_values should be used. Default is c(250, 1000), following the original t-SNE paper. \cr
\code{early_exageration_iter}: \tab Numerical. The number of iterations the early exageration trick should be used. Default 50. \cr
\code{early_exageration_factor}: \tab Numerical. The exageration factor used in the early exacegration phase. Default is 4. \cr
}
}

\section{References}{

\itemize{
\item L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research 9(Nov):2579-2605, 2008.
\item L.J.P. van der Maaten. The t-SNE FAQ \url{https://lvdmaaten.github.io/tsne/}
}
}

\examples{
xx <- make_swiss_roll(150)
tsne_res <- tsne(xx[,1:3])
plot(tsne_res$par, col=xx[,4], pch=16)

}
\seealso{
The \code{\link{tsne_phyloseq}} function provides a nice interface for using t-SNE with microbiota data.
}
