% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_phyloseq.r
\name{tsne_phyloseq}
\alias{tsne_phyloseq}
\title{Dimension reduction of microbiota data with t-SNE}
\usage{
tsne_phyloseq(physeq, distance = "bray", perplexity = 25, dimensions = 2,
  precomputed_distance = NULL, pseudocounts = 1, verbose = 1,
  rng_seed = NULL, philr_options = list(), control = list())
}
\arguments{
\item{physeq}{Phylogenetic sequencing data (\code{\link[phyloseq]{phyloseq-class}}).}

\item{distance}{The dissimilarity measure to use. Either one of \code{bray} (default) \code{unifrac}, \code{wunifrac}, \code{jsd}, \code{dpcoa} or \code{philr}.}

\item{perplexity}{The perplexity. Usually a number between 5 and 50. Default is 25.}

\item{dimensions}{The number of dimensions of the t-SNE embedding. Default is 2, which is usually adequate.}

\item{precomputed_distance}{Matrix or dist object of a precomputed dissimilarity matrix. If you use other ordination techniques or trying different perplexities it will be faster to compute the dissimilarities beforehand.}

\item{pseudocounts}{The number of psuedocounts, which is added to all elements in the otu table. Only applied when distance is philr.}

\item{verbose}{The amount of output during the t-SNE estimation. Can be 0 (no output), 1 (some output, including a progress bar) and 2 (detailed output, mostly usefull for debugging purposes)}

\item{rng_seed}{Provide a seed for generating intiial values for the coordiantes. The final t-SNE result is highly sensitive for starting values, so to get reproducible results you should provide a seed.}

\item{philr_options}{A list of options to use with philr distances. See 'Details'.}

\item{control}{A list of control parameters. See the 'Details' section for \code{\link{tsne}}.}
}
\value{
A list of class tsne_phyloseq.

\tabular{ll}{

\code{distance_matrix} \tab The distance matrix computed using the method
specified in the \code{distance} argument. \cr

\code{sne_distance_matrix} \tab The SNE dissimilarity matrix computed from the
distance_matrix with the perplexity given via the perplexity argument. \cr

\code{names} \tab Samples names from the phyloseq object. \cr

\code{tsne} \tab A list of class \code{tsne} as returned from the \code{\link{tsne}} function. Contains the t-SNE layout and some fit diagnostics,  \cr

}
}
\description{
t-Distributed Stochastic Neighbor Embedding (t-SNE) for microbiota data.
}
\details{
The \code{philr_options} argument can conatin two elements, part.weights and ilr.weights.
The default is 'enorm.x.gm.counts' and 'blw.sqrt', respectively, since these are
used in the philr vignette. See \code{\link[philr]{philr}} for more details.
}
\section{References}{

\itemize{
\item L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research 9(Nov):2579-2605, 2008.
\item L.J.P. van der Maaten. The t-SNE FAQ \url{https://lvdmaaten.github.io/tsne/}
\item Paul J. McMurdie & Susan Holmes phyloseq: An R Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data \url{https://doi.org/10.1371/journal.pone.0061217}
}
}

\examples{
library(phyloseq)
data("GlobalPatterns")
tsne_res <- tsne_phyloseq(GlobalPatterns, distance='wunifrac',
   perplexity = 8, verbose=0, rng_seed = 3901)
plot_tsne_phyloseq(GlobalPatterns, tsne_res,
   color = 'SampleType', title='t-SNE (Weighted UniFrac)')

}
\seealso{
\code{\link{plot_tsne_phyloseq}}, \code{\link{add_tsne_to_sample_data}},
\code{\link{tsne}}.

Other dimension reduction methods for microbiota data is available
via the \code{\link[phyloseq]{ordinate}} function in the phyloseq package.
}
